Update which filename is checked for BinExport
Usually the filename and the exefilename in the BinDiff file is the
same, however I noticed when using BinDiff's export plugin, it names the
binexport file based on the value found in the filename field, not the
exefilename.  Additionally, the program name as gotten by
program.getName() may not match the filename as well.  Using the
"domain file" seems to fix this issue.