Show functions that are only on one side
In the table view, show functions that are only in the primary or
secondary in the table view.  Clicking on ones in the secondary is a bad
idea.  It will throw a null pointer exception.  The on-click handler
should probably pay attention and if the primary is null then do
nothing.